WeaterAppDemo/app/src/main/java/com/hexaware/weatherapp/room/LocationDao.kt
/*
 * Copyright (C) 2017 Google Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.hexaware.weatherapp.room


import androidx.room.Dao
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import com.hexaware.weatherapp.model.LocationTable
import kotlinx.coroutines.flow.Flow


@Dao
interface LocationDao {

    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insert(location: LocationTable)

    @Query("DELETE FROM location_table where latitude=:latLng")
    suspend fun deleteRow(latLng: Double)

    @Query("DELETE FROM location_table ")
    suspend fun deleteAll()


    // The flow always holds/caches latest version of data. Notifies its observers when the
    // data has changed.
    @Query("SELECT * FROM location_table ORDER BY id DESC")
    fun getAllLocation(): Flow<List<LocationTable>>
}
